requirements.txt
pandas==2.2.2 
scikit-learn==1.5.1 
xgboost==2.1.1 
mlflow==2.14.3 
azure-storage-blob==12.19.0 
joblib==1.4.2 
matplotlib==3.9.0 
seaborn==0.13.2 
python-dotenv==1.0.1 
streamlit==1.36.0 
fastapi==0.111.0 
uvicorn==0.30.1 
requests==2.31.0 
